Backend/nelium_air_control/requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.14.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
influxdb-client==1.33.0
